[Feature] CRUD Drink
- Add drink model
- Add drink migration
- Add drink pg service
- Add drink endpoints
- Add drink decoder
- Update drink route in http.go
- Update drink services in main.go

+++ domain/drink.go
@@ -3,6 +3,6 @@ package domain
 // Drink describe drink in system
 type Drink struct {
 	Model
-	Name  string  `json:"name"`
-	Price float64 `json:"price"`
+	Name  string `json:"name"`
+	Price int    `json:"price"`
 }
